% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{nsavi}
\alias{nsavi}
\title{Calculate nSAVI}
\usage{
nsavi(b1, b2, L = 0.5)
}
\arguments{
\item{b1}{value or vector with reflectance in the first (eg. NIR for NDVI) wavelenght}

\item{b2}{value or vector with reflectance in the second (eg. red for NDVI) wavelenght}

\item{L}{soil coefficient. Set to 0.5 by default}
}
\value{
NDVI value
}
\description{
This function caculates the normalized difference vegetation index.
}
